게임 업데이트 시원씨 땡큐요~
게임 업데이트 시원씨 땡큐요~

+++ src/main/java/org/koreait/games/controllers/RequestShadowGame.java
@@ -11,7 +11,7 @@ public class RequestShadowGame implements Serializable {
 
     private List<Long> pokemonCount; // 현재 난이도에 맞게 포켓몬을 넣어주는 List
 
-    private int pokemonCheck; // 상 - 898, 중 - 649, 하 - 386으로 포켓몬 몇명 뽑았는지 체크
+    private Long pokemonCheck; // 상 - 898, 중 - 649, 하 - 386으로 포켓몬 몇명 뽑았는지 체크
     private boolean row; // 난이도 하
     private boolean mid; // 난이도 중
     private boolean high; // 난이도 상
@@ -19,7 +19,6 @@ public class RequestShadowGame implements Serializable {
     private String pokemonName; // 입력한 포켓몬 이름
 
     private int gameCount; // 현재 게임 Count
-    // private Long gamePokemonIndex; // 뽑은 포켓몬 seq
     private Pokemon gamePokemon;
 
     private boolean gameQuizAnswer; // 정답 맞췄는지 못맞췄는지 확인. 이거 지워질 확률 있음.
